pyproject.toml
[project]
name = "claude-hermes"
version = "0.1.0"
description = "Planner-Generator-Evaluator harness for iterative autonomous program development"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
]

[project.scripts]
claude-hermes = "hermes.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hermes"]
